rename gui example and restore gltf-vertex-skinning example
